Clubs Endpoint Completed
